Search component created
